Start terminal and detach
I did not notice this bug, because I had a terminal already open on
every startup but if I have no xfce4-terminal open, Rofi is launching
the terminal and waits for its finish (which does not happens while the
user does not close the terminal), but Rofi window is not closed until
that point, so it seems that Rofi is frozen.

Open xfce4-terminal with `nohup` to detach it from Rofi.